solver: move error handling into its own module
